Update note-taker.js
Adds support for storing different names, rather than just having my name hard coded. The idea for this is that it will prompt you 1 time for your name, and then it stores it in Chrome storage. Not entering a name will cause this bookmarklet to not function, which I imagine is fine since the entire point of this bookmarklet is to leave notes.